refactor/point: CQRS 패턴 적용 및 코드 컨벤션 정리 (Apply CQRS pattern and format code conventions)
