--- src/main/java/org/example/lastcall/domain/point/service/PointServiceApi.java
@@ -1,10 +1,0 @@
-package org.example.lastcall.domain.point.service;
-
-public interface PointServiceApi {
-
-    void validateSufficientPoints(Long userId, Long requiredAmount);
-
-    void updateDepositPoint(Long auctionId, Long bidId, Long bidAmount, Long userId);
-
-    void depositToSettlement(Long userId, Long auctionId, Long amount);
-}
