+++ src/main/java/org/example/lastcall/domain/point/service/command/PointCommandServiceApi.java
@@ -1,0 +1,7 @@
+package org.example.lastcall.domain.point.service.command;
+
+public interface PointCommandServiceApi {
+	void updateDepositPoint(Long auctionId, Long bidId, Long bidAmount, Long userId);
+
+	void depositToSettlement(Long userId, Long auctionId, Long amount);
+}
