+++ src/main/java/org/example/lastcall/domain/point/service/query/PointQueryServiceApi.java
@@ -1,0 +1,5 @@
+package org.example.lastcall.domain.point.service.query;
+
+public interface PointQueryServiceApi {
+	void validateSufficientPoints(Long userId, Long requiredAmount);
+}
